Use test description instead of suite description to prevent NPE in gradle junit handler
The gradle handler cannot resolve the method name on the test related callbacks that are fired for the scenario. This later causes a NPE when trying to transform the binary test output into an XML report.
By providing a valid method name the handler can successfully digest the test output.